+++ QuEST/QuEST_internal.h
@@ -15,7 +15,7 @@ extern "C" {
 
 unsigned long int hashString(char *str);
 
-REAL getVectorMagnitude(Vector vec);
+qreal getVectorMagnitude(Vector vec);
 
 Complex getConjugateScalar(Complex scalar);
 
@@ -23,11 +23,11 @@ ComplexMatrix2 getConjugateMatrix(ComplexMatrix2 matr);
 
 void ensureIndsIncrease(int* ind1, int* ind2);
 
-void getComplexPairFromRotation(REAL angle, Vector axis, Complex* alpha, Complex* beta);
+void getComplexPairFromRotation(qreal angle, Vector axis, Complex* alpha, Complex* beta);
 
-void getZYZRotAnglesFromComplexPair(Complex alpha, Complex beta, REAL* rz2, REAL* ry, REAL* rz1);
+void getZYZRotAnglesFromComplexPair(Complex alpha, Complex beta, qreal* rz2, qreal* ry, qreal* rz1);
 
-void getComplexPairAndPhaseFromUnitary(ComplexMatrix2 u, Complex* alpha, Complex* beta, REAL* globalPhase);
+void getComplexPairAndPhaseFromUnitary(ComplexMatrix2 u, Complex* alpha, Complex* beta, qreal* globalPhase);
 
 void shiftIndices(int* indices, int numIndices, int shift);
