+++ QuEST/QuEST_qasm.h
@@ -38,33 +38,33 @@ void qasm_stopRecording(Qureg qureg);
 
 void qasm_recordGate(Qureg qureg, TargetGate gate, int targetQubit);
 
-void qasm_recordParamGate(Qureg qureg, TargetGate gate, int targetQubit, REAL param);
+void qasm_recordParamGate(Qureg qureg, TargetGate gate, int targetQubit, qreal param);
 
 void qasm_recordCompactUnitary(Qureg qureg, Complex alpha, Complex beta, int targetQubit);
 
 void qasm_recordUnitary(Qureg qureg, ComplexMatrix2 u, int targetQubit);
 
-void qasm_recordAxisRotation(Qureg qureg, REAL angle, Vector axis, const int targetQubit);
+void qasm_recordAxisRotation(Qureg qureg, qreal angle, Vector axis, const int targetQubit);
 
 void qasm_recordControlledGate(Qureg qureg, TargetGate gate, int controlQubit, int targetQubit);
 
-void qasm_recordControlledParamGate(Qureg qureg, TargetGate gate, int controlQubit, int targetQubit, REAL param);
+void qasm_recordControlledParamGate(Qureg qureg, TargetGate gate, int controlQubit, int targetQubit, qreal param);
 
 void qasm_recordControlledCompactUnitary(Qureg qureg, Complex alpha, Complex beta, int controlQubit, int targetQubit);
 
 void qasm_recordControlledUnitary(Qureg qureg, ComplexMatrix2 u, int controlQubit, int targetQubit);
 
-void qasm_recordControlledAxisRotation(Qureg qureg, REAL angle, Vector axis, int controlQubit, int targetQubit);
+void qasm_recordControlledAxisRotation(Qureg qureg, qreal angle, Vector axis, int controlQubit, int targetQubit);
 
 void qasm_recordMultiControlledGate(Qureg qureg,
 TargetGate gate, int* controlQubits, const int numControlQubits, const int targetQubit);
 
-void qasm_recordMultiControlledParamGate(Qureg qureg, TargetGate gate, int* controlQubits, const int numControlQubits, const int targetQubit, REAL param);
+void qasm_recordMultiControlledParamGate(Qureg qureg, TargetGate gate, int* controlQubits, const int numControlQubits, const int targetQubit, qreal param);
 
 void qasm_recordMultiControlledUnitary(Qureg qureg, ComplexMatrix2 u, int* controlQubits, const int numControlQubits, const int targetQubit);
 
 /* not actually used. D'oh!
-void qasm_recordMultiControlledAxisRotation(Qureg qureg, REAL angle, Vector axis, int* controlQubits, const int numControlQubits, const int targetQubit);\
+void qasm_recordMultiControlledAxisRotation(Qureg qureg, qreal angle, Vector axis, int* controlQubits, const int numControlQubits, const int targetQubit);\
 */
 
 void qasm_recordMeasurement(Qureg qureg, const int measureQubit);
